Drop _with_reply in functions used for method invocation
Strictly speaking, it doesn't really make sense to speak of a "reply"
when talking about method calls (it makes sense to talk about the
"result" though). It does make sense to speak of replies in a
message-based API but that's not what we're exposing here.

And if the caller is not interested in the reply, he can just pass
NULL for the GAsyncReadyCallback.