Introduce proxy default timeout
Some D-Bus services can take ages to generate a reply. One example of
this is DeviceKit-disks - if you do FilesystemCreate() the method call
can be pending for hours, even days, while waiting for mkfs(8) to do
it's job.

With tools, such as eggdbus-binding-tool, generating proxy classes it
would be nice to avoid a @timeout_msec parameter. This can be avoided
if the generated proxy always passes -1 for the timeout and GDBusProxy
interprets this as using a timeout configurable for the proxy
instance. With this patch, the user can simply do

 g_dbus_proxy_set_timeout (G_DBUS_PROXY (my_disk_proxy), GMAX_INT);

to force no timeout.